fix(frontend): kanban task fields update reactively from detail panel
KanbanView held allTasks as local component state populated once on
mount. Detail-panel motif/text/etc. edits PATCH the API and refresh
plan/backlog stores, but never touched the kanban view's stale copy —
so changing a task's motif updated the sidebar in PlanView but not in
the Kanban column.

Hoist allTasks + fetchAllTasks into kanbanStore. Add applyTaskPatch
that merges the patch into the in-memory copy. TaskDetailPanel.patchTask
calls it after a successful PATCH so any view bound to kanbanStore.allTasks
reflects the change immediately. No-op when the kanban list isn't loaded
(detail panel may be open without ever visiting kanban this session).
