+++ frontend/src/stores/kanban.ts
@@ -1,6 +1,7 @@
 import { defineStore } from 'pinia'
 import { ref } from 'vue'
 import { api } from '../lib/api'
+import type { Task } from './plan'
 
 export interface KanbanColumn {
   id: string
@@ -12,9 +13,16 @@ export interface KanbanColumn {
   created_by: string | null
 }
 
+export interface KanbanTask extends Task {
+  plan_date: string | null
+  anchor_id: string | null
+}
+
 export const useKanbanStore = defineStore('kanban', () => {
   const columns = ref<KanbanColumn[]>([])
+  const allTasks = ref<KanbanTask[]>([])
   const loading = ref(false)
+  const tasksLoading = ref(false)
   const error = ref<string | null>(null)
 
   async function fetchColumns() {
@@ -34,5 +42,38 @@ export const useKanbanStore = defineStore('kanban', () => {
     }
   }
 
-  return { columns, loading, error, fetchColumns }
+  async function fetchAllTasks() {
+    tasksLoading.value = true
+    try {
+      const resp = await api('/api/tasks/all')
+      if (!resp.ok) throw new Error(`${resp.status}`)
+      allTasks.value = await resp.json()
+    } catch (e) {
+      console.error('fetchAllTasks error:', e)
+    } finally {
+      tasksLoading.value = false
+    }
+  }
+
+  /**
+   * Merge a partial PATCH into the local task copy so views holding
+   * `allTasks` re-render immediately after a detail-panel edit.
+   * No-op if the task isn't in the local list (e.g. kanban not loaded yet).
+   */
+  function applyTaskPatch(taskId: string, patch: Partial<KanbanTask>) {
+    const i = allTasks.value.findIndex(t => t.id === taskId)
+    if (i === -1) return
+    allTasks.value[i] = { ...allTasks.value[i], ...patch }
+  }
+
+  return {
+    columns,
+    allTasks,
+    loading,
+    tasksLoading,
+    error,
+    fetchColumns,
+    fetchAllTasks,
+    applyTaskPatch,
+  }
 })
